requirements/base.txt
amqp==5.2.0
asgiref==3.8.1
attrs==23.2.0
beautifulsoup4==4.12.3
billiard==4.2.0
bs4==0.0.2
celery==5.4.0
certifi==2024.7.4
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.3
Django==5.0.6
django-celery-beat==2.6.0
django-celery-results==2.5.1
django-filter==24.2
django-timezone-field==6.1.0
djangorestframework==3.15.2
h11==0.14.0
idna==3.7
kombu==5.3.7
Markdown==3.6
outcome==1.3.0.post0
pillow==10.4.0
prompt_toolkit==3.0.47
PySocks==1.7.1
python-crontab==3.2.0
python-dateutil==2.9.0.post0
redis==5.0.7
selenium==4.22.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
sqlparse==0.5.0
trio==0.26.0
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
vine==5.1.0
wcwidth==0.2.13
websocket-client==1.8.0
wsproto==1.2.0